Removes `rsaObjects` files as no longer needed
