--- src/rsa/rsaObjects.cpp
@@ -1,15 +1,0 @@
-/*
- * Copyright 2023-2024 The Gestalt Project Authors. All Rights Reserved.
- *
- * Licensed under the MIT License. See the file LICENSE for the full text.
- */
-
-/*
- * rsaObjects.cpp
- *
- */
-
-#include <iostream>
-#include <cmath>
-
-#include "rsaObjects.h"
